Move function in run.py to generate_string.py
